+++ src-tauri/src/frontend/mod.rs
@@ -1,1 +1,4 @@
-pub mod main;
+/// functions used primarily by the backend
+pub mod functions;
+/// tauri initialization and command declarations
+pub mod tauri;
