+++ src-tauri/src/frontend/tauri.rs
@@ -1,7 +1,8 @@
-use log::{debug, error, info, warn};
-use tauri::api::cli::ArgData;
+use log::error;
 
-use crate::backend::{config_file::Config, utils::{self, generic::{get_config, update_config}, scanner_utils}};
+use crate::backend::{config_file::Config, utils::{self, generic::{get_config, update_config}}};
+
+use super::functions::{cli_dbupdate, cli_gui, cli_scanner, not_implemented, print_data};
 
 pub fn init_tauri() {
     // Builds the Tauri connection
@@ -56,86 +57,10 @@ pub fn init_tauri() {
         .expect("error while running tauri application");
 }
 
-#[cfg(all(not(debug_assertions), windows))]
-fn remove_windows_console() {
-    unsafe {
-        windows_sys::Win32::System::Console::FreeConsole();
-    }
-}
-
-// Basically prints the given data with \n and \t correctly formatted
-fn print_data(app: tauri::AppHandle, data: &ArgData) {
-    if let Some(json_str) = data.value.as_str() {
-        let unescaped_str = json_str.replace("\\n", "\n").replace("\\t", "\t");
-        debug!("{}", unescaped_str);
-        app.exit(0);
-    } else {
-        // Handle the case where data.value is not a string
-        error!("data.value is not a string");
-        app.exit(1);
-    }
-}
-
-// If a command is not yet implemented
-fn not_implemented(app: tauri::AppHandle) {
-    warn!("Function not implemented yet");
-    app.exit(2);
-}
-
-// Starts the GUI without attaching a CLI
-fn cli_gui(app: tauri::AppHandle) -> Result<(), tauri::Error> {
-    debug!("Showing GUI...");
-    #[cfg(all(not(debug_assertions), windows))]
-    remove_windows_console();
-    tauri::WindowBuilder::new(&app, "raspirus", tauri::WindowUrl::App("index.html".into()))
-        .title("Raspirus")
-        .inner_size(800., 480.)
-        .resizable(true)
-        .build()?;
-    debug!("This won't show on Windows release builds");
-    Ok(())
-}
-
-// Starts the scanner on the CLI
-fn cli_scanner(app: tauri::AppHandle, data: &ArgData) {
-    if let Some(json_str) = data.value.as_str() {
-        let unescaped_str = json_str.replace("\\n", "\n").replace("\\t", "\t");
-        debug!("Data provided: {}", unescaped_str);
-        match scanner_utils::start_scanner(None, unescaped_str) {
-            Ok(res) => {
-                info!("Result: {res}");
-                app.exit(0);
-            }
-            Err(err) => {
-                error!("Error: {err}");
-                app.exit(-1);
-            }
-        }
-    } else {
-        // Handle the case where data.value is not a string
-        error!("data.value is not a string");
-        app.exit(-1);
-    }
-}
-
-// Updates the DB over the CLI
-fn cli_dbupdate(app: tauri::AppHandle) {
-    match utils::update_utils::update(None) {
-        Ok(res) => {
-            info!("Result: {res}");
-            app.exit(0);
-        }
-        Err(err) => {
-            error!("Error: {err}");
-            app.exit(-1);
-        }
-    }
-}
-
 // Starts the scanner over the GUI
 #[tauri::command]
-async fn start_scanner(window: tauri::Window, path: String) -> Result<String, String> {
-    tokio::task::spawn_blocking(|| scanner_utils::start_scanner(Some(window), path))
+pub async fn start_scanner(window: tauri::Window, path: String) -> Result<String, String> {
+    tokio::task::spawn_blocking(|| utils::scanner_utils::start_scanner(Some(window), path))
         .await
         .map_err(|err| err.to_string())?
 }
@@ -143,20 +68,20 @@ async fn start_scanner(window: tauri::Window, path: String) -> Result<String, St
 // Checks if we are currently on a Raspberry Pi,
 // because a couple options are not supported on that device and will be disabled on the GUI
 #[tauri::command]
-async fn check_raspberry() -> Result<bool, String> {
+pub async fn check_raspberry() -> Result<bool, String> {
     Ok(std::env::consts::ARCH == "arm")
 }
 
 // Updates the database over the GUi
 #[tauri::command]
-async fn update_database(window: tauri::Window) -> Result<String, String> {
+pub async fn update_database(window: tauri::Window) -> Result<String, String> {
     tokio::task::spawn_blocking(|| utils::update_utils::update(Some(window)))
         .await
         .map_err(|err| err.to_string())?
 }
 
 #[tauri::command]
-async fn patch(patchfile: String) -> Result<(), String> {
+pub async fn patch(patchfile: String) -> Result<(), String> {
     tokio::task::spawn_blocking(move || utils::update_utils::patch(&patchfile))
         .await
         .map_err(|err| err.to_string())?
@@ -165,13 +90,13 @@ async fn patch(patchfile: String) -> Result<(), String> {
 
 // Returns a vector of all attached removable storage drives (USB) -> Unnecessary for the CLI
 #[tauri::command]
-async fn list_usb_drives() -> Result<String, String> {
+pub async fn list_usb_drives() -> Result<String, String> {
     utils::usb_utils::list_usb_drives().await
 }
 
 // Creates the config from the GUI
 #[tauri::command]
-fn save_config_fe(contents: Option<String>) -> Result<(), String> {
+pub fn save_config_fe(contents: Option<String>) -> Result<(), String> {
     let mut config = serde_json::from_str::<Config>(&contents.ok_or("Json was none".to_owned())?)
         .map_err(|err| err.to_string())?;
     config.paths = get_config().paths;
@@ -179,13 +104,13 @@ fn save_config_fe(contents: Option<String>) -> Result<(), String> {
 }
 
 #[tauri::command]
-fn load_config_fe() -> Result<String, String> {
+pub fn load_config_fe() -> Result<String, String> {
     serde_json::to_string(&get_config())
         .map_err(|err| format!("Failed to convert config to json: {err}"))
 }
 
 #[tauri::command]
-async fn check_update() -> Result<bool, String> {
+pub async fn check_update() -> Result<bool, String> {
     tokio::task::spawn_blocking(utils::update_utils::check_update_necessary)
         .await
         .map_err(|err| err.to_string())?
@@ -193,7 +118,7 @@ async fn check_update() -> Result<bool, String> {
 }
 
 #[tauri::command]
-async fn download_logs() -> Result<String, String> {
+pub async fn download_logs() -> Result<String, String> {
     let log_dir = get_config()
         .paths
         .ok_or("No paths set. Is config initialized?".to_owned())?
